src/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==3.0.2
Django==4.1.1
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
msgpack==1.0.5
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.1
packaging==21.3
pandas==1.5.3
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
zope.interface==6.0
